+++ packages/outstatic/src/components/ui/outstatic/navigation-config.schema.ts
@@ -1,7 +1,7 @@
-import { z } from 'zod';
+import { z } from 'zod/v4'
 
 const RouteMatchingEnd = z
-  .union([z.boolean(), z.function().args(z.string()).returns(z.boolean())])
+  .union([z.boolean(), z.custom<(input: string) => boolean>()])
   .default(false)
   .optional();
 
@@ -37,19 +37,17 @@ const RouteGroup = z.object({
   label: z.string(),
   collapsible: z.boolean().optional(),
   collapsed: z.boolean().optional(),
-  children: z.array(RouteChild),
+  Icon: z.custom<React.ReactNode>().optional(),
+  get children(): z.ZodUnion<[z.ZodArray<typeof RouteChild>, z.ZodArray<typeof RouteGroup>]> {
+    return z.union([z.array(RouteChild), z.array(RouteGroup)])
+  },
   renderAction: z.custom<React.ReactNode>().optional(),
   badge: z.custom<React.ReactNode>().optional(),
   dialog: z.custom<React.ReactNode>().optional(),
   newTab: z.boolean().default(false).optional(),
-});
+})
 
 export const NavigationConfigSchema = z.object({
-  style: z.enum(['custom', 'sidebar', 'header']).default('sidebar'),
-  sidebarCollapsed: z
-    .union([z.boolean(), z.enum(['false', 'true'])])
-    .default(true)
-    .transform(val => typeof val === 'string' ? val === 'true' : val),
-  sidebarCollapsedStyle: z.enum(['offcanvas', 'icon', 'none']).default('icon'),
-  routes: z.array(z.union([RouteGroup, Divider])),
-});
+  // @ts-ignore
+  routes: z.array(z.union([RouteGroup, Divider]))
+})